Update README.md
After rereading this document, I am able to see points that I could have offered more detail to the read.  By adding more detail and expounding on each point the reader would be better able to create a visual and understanding of the points I made. The organization of the headings is one way that does help the reader follow the main points of the document.  If it were one long reading without the headings it would be more difficult to follow along. At the end of the reading there could be more offered for a looking ahead; however, I think that there is a sense of looking ahead by mentioning the safety of the horse and trainer.  This would leave the reader with enough curiosity to want to do more research. 
@matthewjmiller